Add __penzai_repr__ extension method and repr_lib helpers.
The __penzai_repr__ method can be used to add a handler for a custom
type without modifying the global handler list. This is designed to make
it easier to extend treescope to support custom types.

Also adds a high-level helper library `repr_lib` to make it easier to
implement __penzai_repr__.

PiperOrigin-RevId: 637942807