Add LayerStack combinator backed by jax.lax.scan.
This CL implements a LayerStack combinator, which scans over a named axis of its child
layers when called. State variables inside a LayerStack must be annotated with their
sharing behavior. Parameters and leaf values are either scanned over or replicated
depending on whether they have the particular named axis that identifies the stack.

Also adds a utility to named axes to simplify scanning, and adapts the Transformer
implementation to support layer scan.

PiperOrigin-RevId: 646095346