Fix NamedArray RHS bin-op promotion; remove implicit array conversion.
Due to Python's resolution rules for binary operations like + and -,
whichever object is on the left-hand side gets priority in implementing
overloads, via e.g. `__add__` or `__sub__`. The right-hand side object is
only consulted if these methods return NotImplemented.

When a Numpy or JAX array is the LHS of a binary operation expression,
the implementation currently tries to convert the NamedArray to a
Numpy or JAX array using `__array__` and `__jax_array__`, which will fail
if the named array has named axes. This prevents `__radd__` or `__rsub__`
from being called on the NamedArray (which then correctly uses nmap).

This change fixes the issue by removing the automatic `__array__` and
`__jax_array__` conversion methods; `unwrap` must now be used instead.
(In the future, it may be possible to add this back using the
`__array_priority__` attribute, which Numpy checks but JAX does not.)

Fixes #35.

PiperOrigin-RevId: 646097598