Add basic `.at[...].set(...)` syntax support to named arrays.
This change adds support for `.at[...].set(...)` syntax on Penzai's
NamedArray and NamedArrayView. Similar to the other array methods,
the semantics of `.at[...].set(...)` are defined based on `nmap`,
such that

```
arr.at[a].set(b)
```

is equivalent to

```
nmap(lambda arr, a, b: arr.at[a].set(b))(arr, a, b)
```

Currently, only updates of positional axes are supported, so named axes must be
`untag`-ged before they can be indexed in this manner. Indexing or
updating with other named arrays is allowed and broadcasts according
to ordinary `nmap` semantics.

PiperOrigin-RevId: 646103893