Fix HTML encapsulation script execution bug
In certain circumstances, the new enapsulation system can fail to find the
script to execute. This can be triggered by using treescope inside an iframe's
srcdoc attribute, and may occur in other cases as well. This appears to be
due to the custom component being connected before its child script is parsed.

This change fixes the issue by detecting the missing script tag and waiting for it
to be loaded before executing it.

PiperOrigin-RevId: 652963362