Refactor treescope to mostly not depend on Penzai core.
This change moves around parts of treescope and Penzai to reduce the dependencies
of treescope on the rest of Penzai core. This is the first step toward allowing
treescope to be installed and imported separately.

To support this, the IPython integration now looks for a special method
`__penzai_root_repr__` that allows types to customize their root representations.
Additionally, the `context` and `dataclass_util` modules have been moved to treescope
instead of Penzai core.

PiperOrigin-RevId: 653401565