Remove dependency on Equinox, change `Selection.at` API.
This change re-implements the necessary logic for `Selection.at`
directly, allowing Penzai to drop its dependency on Equinox.
To simplify the logic, `Selection.at` now takes an additional argument
`multiple`, which determines whether or not the accessor function is
expected to return a single value or a collection of values in the tree.

PiperOrigin-RevId: 656556966